cdk/requirements.txt
aws-cdk-lib
constructs>=10.0.0,<11.0.0
requests
aws-solutions-constructs.aws-lambda-eventbridge
boto3
